+++ lib/eligible.rb
@@ -44,6 +44,7 @@ require 'eligible/icd'
 # New REST API Endpoints
 require 'eligible/v1_0/file_object'
 require 'eligible/v1_0/patient_statement'
+require 'eligible/v1_0/patient_statement_service_line'
 
 # Errors
 require 'eligible/errors/eligible_error'